+++ Sources/EmojiKit/Views/EmojiGrid.swift
@@ -12,16 +12,20 @@ import SwiftUI
 ///
 /// You can use an ``EmojiScrollGrid`` to wrap the grid in a
 /// `ScrollView` that auto-scrolls to any new selections and
-/// injects a `geometryProxy` to enable arrow-key selections.
+/// that automatically enables arrow-key selections.
 ///
 /// See the <doc:Views-Article> article for full information
 /// on how to use these grids.
 public struct EmojiGrid<ItemView: View, SectionView: View>: View {
     
-    /// Create an emoji grid with multiple category sections.
+    /// Create an emoji grid.
+    ///
+    /// If you provide a list of `emojis`, that list will be
+    /// listed in the grid instead of the `categories`.
     ///
     /// - Parameters:
     ///   - axis: The grid axis, by default `.vertical`.
+    ///   - emojis: A custom emoji collection to list, if any.
     ///   - categories: The categories to list, by default `.all`.
     ///   - query: The search query to apply, if any.
     ///   - selection: The current grid selection, if any.
@@ -32,18 +36,22 @@ public struct EmojiGrid<ItemView: View, SectionView: View>: View {
     ///   - item: A grid item view builder.
     public init(
         axis: Axis.Set = .vertical,
+        emojis: [Emoji] = [],
         categories: [EmojiCategory] = .all,
         query: String = "",
         selection: Binding<Emoji.GridSelection> = .constant(.init()),
         frequentEmojiProvider: (any FrequentEmojiProvider)? = MostRecentEmojiProvider(),
         geometryProxy: GeometryProxy? = nil,
-        action: @escaping EmojiAction = { _ in },
-        @ViewBuilder section: @escaping SectionViewBuilder,
-        @ViewBuilder item: @escaping ItemViewBuilder
+        action: @escaping (Emoji) -> Void = { _ in },
+        @ViewBuilder section: @escaping (Emoji.GridSectionParameters) -> SectionView,
+        @ViewBuilder item: @escaping (Emoji.GridItemParameters) -> ItemView
     ) {
+        let emojiChars = emojis.map { $0.char }.joined()
+        let emojiCat = EmojiCategory.custom(id: "", name: "", emojis: emojiChars, iconName: "")
+        let emojiCategories: [EmojiCategory]? = emojis.isEmpty ? nil : [emojiCat]
+        let searchCategories: [EmojiCategory]? = query.isEmpty ? nil : [.search(query: query)]
         self.axis = axis
-        self.categories = query.isEmpty ? categories : [.search(query: query)]
-
+        self.categories = searchCategories ?? emojiCategories ?? categories
         self.query = query
         self.frequentEmojiProvider = frequentEmojiProvider
         self.geometryProxy = geometryProxy
@@ -53,53 +61,14 @@ public struct EmojiGrid<ItemView: View, SectionView: View>: View {
         self._selection = selection
     }
     
-    /// Create an emoji grid with a single section.
-    ///
-    /// - Parameters:
-    ///   - axis: The grid axis, by default `.vertical`.
-    ///   - emojis: The emojis to list.
-    ///   - query: The search query to apply, if any.
-    ///   - selection: The current grid selection, if any.
-    ///   - frequentEmojiProvider: The ``FrequentEmojiProvider`` to use, if any.
-    ///   - geometryProxy: An optional geometry proxy, required to perform arrow/move-based navigation.
-    ///   - action: An action to trigger when an emoji is tapped or picked.
-    ///   - item: A grid item view builder.
-    public init(
-        axis: Axis.Set = .vertical,
-        emojis: [Emoji],
-        query: String = "",
-        selection: Binding<Emoji.GridSelection> = .constant(.init()),
-        frequentEmojiProvider: (any FrequentEmojiProvider)? = MostRecentEmojiProvider(),
-        geometryProxy: GeometryProxy? = nil,
-        action: @escaping EmojiAction = { _ in },
-        @ViewBuilder item: @escaping ItemViewBuilder
-    ) where SectionView == Emoji.GridSectionTitle {
-        let chars = emojis.map { $0.char }.joined()
-        self.init(
-            axis: axis,
-            categories: [.custom(id: "", name: "", emojis: chars, iconName: "")],
-            query: query,
-            selection: selection,
-            frequentEmojiProvider: frequentEmojiProvider,
-            geometryProxy: geometryProxy,
-            action: action,
-            section: { $0.view },
-            item: item
-        )
-    }
-    
-    public typealias EmojiAction = (Emoji) -> Void
-    public typealias ItemViewBuilder = (Emoji.GridItemParameters) -> ItemView
-    public typealias SectionViewBuilder = (Emoji.GridSectionParameters) -> SectionView
-    
     private let axis: Axis.Set
     private let categories: [EmojiCategory]
     private let query: String
     private let frequentEmojiProvider: (any FrequentEmojiProvider)?
     private let geometryProxy: GeometryProxy?
-    private let action: EmojiAction
-    private let section: SectionViewBuilder
-    private let item: ItemViewBuilder
+    private let action: (Emoji) -> Void
+    private let section: (Emoji.GridSectionParameters) -> SectionView
+    private let item: (Emoji.GridItemParameters) -> ItemView
 
     @Binding
     private var selection: Emoji.GridSelection
@@ -258,38 +227,6 @@ private extension EmojiGrid {
     }
 }
 
-public extension Emoji {
-    
-    /// This struct defines item view builder parameters.
-    struct GridItemParameters {
-        
-        /// The emoji to present.
-        public let emoji: Emoji
-        
-        /// The category of the emoji.
-        public let category: EmojiCategory
-        
-        /// The category index of the emoji.
-        public let categoryIndex: Int
-        
-        /// Whether or not the emoji is selected.
-        public let isSelected: Bool
-        
-        /// The standard grid item view.
-        public let view: Emoji.GridItem
-    }
-    
-    /// This struct defines section view builder parameters.
-    struct GridSectionParameters {
-        
-        /// The category that is to be presented.
-        public let category: EmojiCategory
-        
-        /// The standard grid item view.
-        public let view: Emoji.GridSectionTitle
-    }
-}
-
 private extension EmojiGrid {
 
     @ViewBuilder
@@ -408,20 +345,12 @@ private extension EmojiGrid {
     }
 }
 
-struct EmojiGridPreviewButtonStyle: ButtonStyle {
-    
-    func makeBody(configuration: Configuration) -> some View {
-        configuration.label
-            .opacity(configuration.isPressed ? 0.5 : 1)
-    }
-}
-
 #Preview {
     
     struct Preview: View {
 
         @State
-        var query = "lollipop"
+        var query = ""
 
         @State
         var selection = Emoji.GridSelection(emoji: .init("👼"), category: .smileysAndPeople)
@@ -435,16 +364,8 @@ struct EmojiGridPreviewButtonStyle: ButtonStyle {
                         axis: axis,
                         query: query,
                         selection: $selection,
-                        // frequentEmojiProvider: provider,
                         section: { $0.view },
-                        item: { params in
-                            Button {
-                                select(params.emoji, in: params.category)
-                            } label: {
-                                params.view
-                            }
-                            .buttonStyle(EmojiGridPreviewButtonStyle())
-                        }
+                        item: { $0.view }
                     )
                     .padding(5)
                 }
@@ -455,13 +376,6 @@ struct EmojiGridPreviewButtonStyle: ButtonStyle {
                     proxy.scrollTo(selection)
                 }
             }
-        }
-        
-        func select(
-            _ emoji: Emoji,
-            in cat: EmojiCategory
-        ) {
-            selection = .init(emoji: emoji, category: cat)
         }
         
         var body: some View {