+++ Sources/EmojiKit/Views/EmojiScrollGrid.swift
@@ -8,86 +8,67 @@
 
 import SwiftUI
 
-/// This scroll grid wraps an ``EmojiGrid`` in a `ScrollView`
-/// and automatically scrolls to the current `selection`.
+/// This scroll grid wraps an ``EmojiGrid``, auto-scrolls to
+/// the `selection` and auto-enables keyboard support.
 ///
 /// See the <doc:Views-Article> article for full information
 /// on how to use these grids.
 public struct EmojiScrollGrid<ItemView: View, SectionView: View>: View {
     
-    /// Create an emoji grid with multiple category sections.
+    /// Create an emoji scroll grid.
+    ///
+    /// If you provide a list of `emojis`, that list will be
+    /// listed in the grid instead of the `categories`.
     ///
     /// - Parameters:
     ///   - axis: The grid axis, by default `.vertical`.
+    ///   - emojis: A custom emoji collection to list, if any.
     ///   - categories: The categories to list, by default `.all`.
     ///   - query: The search query to apply, if any.
     ///   - selection: The current grid selection, if any.
     ///   - frequentEmojiProvider: The ``FrequentEmojiProvider`` to use, by default a ``MostRecentEmojiProvider``.
+    ///   - geometryProxy: An optional geometry proxy, required to perform arrow/move-based navigation.
     ///   - action: An action to trigger when an emoji is tapped or picked.
     ///   - section: A grid section title view builder.
     ///   - item: A grid item view builder.
     public init(
         axis: Axis.Set = .vertical,
+        emojis: [Emoji] = [],
         categories: [EmojiCategory] = .all,
         query: String = "",
         selection: Binding<Emoji.GridSelection> = .constant(.init()),
         frequentEmojiProvider: (any FrequentEmojiProvider)? = MostRecentEmojiProvider(),
-        action: @escaping EmojiAction = { _ in },
-        @ViewBuilder section: @escaping SectionViewBuilder,
-        @ViewBuilder item: @escaping ItemViewBuilder
+        geometryProxy: GeometryProxy? = nil,
+        action: @escaping (Emoji) -> Void = { _ in },
+        @ViewBuilder section: @escaping (Emoji.GridSectionParameters) -> SectionView,
+        @ViewBuilder item: @escaping (Emoji.GridItemParameters) -> ItemView
     ) {
+        let emojiChars = emojis.map { $0.char }.joined()
+        let emojiCat = EmojiCategory.custom(id: "", name: "", emojis: emojiChars, iconName: "")
+        let emojiCategories: [EmojiCategory]? = emojis.isEmpty ? nil : [emojiCat]
+        let searchCategories: [EmojiCategory]? = query.isEmpty ? nil : [.search(query: query)]
         self.axis = axis
-        self.categories = categories
+        self.emojis = emojis
+        self.categories = searchCategories ?? emojiCategories ?? categories
         self.query = query
         self.frequentEmojiProvider = frequentEmojiProvider
+        self.geometryProxy = geometryProxy
         self.action = action
         self.section = section
         self.item = item
         self._selection = selection
     }
-    
-    /// Create an emoji grid with a single section.
-    ///
-    /// - Parameters:
-    ///   - axis: The grid axis, by default `.vertical`.
-    ///   - emojis: The emojis to list.
-    ///   - query: The search query to apply, if any.
-    ///   - selection: The current grid selection, if any.
-    ///   - frequentEmojiProvider: The ``FrequentEmojiProvider`` to use, if any.
-    ///   - action: An action to trigger when an emoji is tapped or picked.
-    ///   - item: A grid item view builder.
-    public init(
-        axis: Axis.Set = .vertical,
-        emojis: [Emoji],
-        query: String = "",
-        selection: Binding<Emoji.GridSelection> = .constant(.init()),
-        frequentEmojiProvider: (any FrequentEmojiProvider)? = MostRecentEmojiProvider(),
-        action: @escaping EmojiAction = { _ in },
-        @ViewBuilder item: @escaping ItemViewBuilder
-    ) where SectionView == Emoji.GridSectionTitle {
-        let chars = emojis.map { $0.char }.joined()
-        self.axis = axis
-        self.query = query
-        self.categories = [.custom(id: "", name: "", emojis: chars, iconName: "")]
-        self.frequentEmojiProvider = frequentEmojiProvider
-        self.action = action
-        self.section = { $0.view }
-        self.item = item
-        self._selection = selection
-    }
-    
-    public typealias EmojiAction = (Emoji) -> Void
-    public typealias ItemViewBuilder = (Emoji.GridItemParameters) -> ItemView
-    public typealias SectionViewBuilder = (Emoji.GridSectionParameters) -> SectionView
-    
+
     private let axis: Axis.Set
+    private let emojis: [Emoji]
     private let categories: [EmojiCategory]
     private let query: String
     private let frequentEmojiProvider: (any FrequentEmojiProvider)?
-    private let action: EmojiAction
-    private let section: SectionViewBuilder
-    private let item: ItemViewBuilder
-    
+    private let geometryProxy: GeometryProxy?
+    private let action: (Emoji) -> Void
+    private let section: (Emoji.GridSectionParameters) -> SectionView
+    private let item: (Emoji.GridItemParameters) -> ItemView
+
     @Binding
     private var selection: Emoji.GridSelection
     
@@ -100,6 +81,7 @@ public struct EmojiScrollGrid<ItemView: View, SectionView: View>: View {
                 ScrollView(axis) {
                     EmojiGrid(
                         axis: axis,
+                        emojis: emojis,
                         categories: categories,
                         query: query,
                         selection: $selection,
@@ -140,14 +122,7 @@ public struct EmojiScrollGrid<ItemView: View, SectionView: View>: View {
                 selection: $selection,
                 // frequentEmojiProvider: provider,
                 section: { $0.view },
-                item: { params in
-                    Button {
-                        select(params.emoji, cat: params.category)
-                    } label: {
-                        params.view
-                    }
-                    .buttonStyle(EmojiGridPreviewButtonStyle())
-                }
+                item: { $0.view }
             )
         }